add nodes_with_system_id from secret watch
